Conflicts:
	src/gui/Source/Views/AnalysisView.mm
	src/gui/Source/Views/ModelView.mm
	src/gui/Source/Views/NotebookView.mm
	src/gui/Source/Views/RbView.mm
	src/gui/Source/Views/SimulateView.mm
	src/gui/Source/Views/TreeSetView.mm
	src/gui/main.mm

+++ src/core/moves/characterhistory/NodeUniformizationSampleProposal.h
@@ -403,7 +403,7 @@ void RevBayesCore::NodeUniformizationSampleProposal<charType, treeType>::sampleN
             
             double u = GLOBAL_RNG->uniform01() * gSum;
             unsigned int s = 0;
-            for (size_t i = 0; i < numStates; i++)
+            for (unsigned i = 0; i < numStates; i++)
             {
                 u -= g[i];
                 if (u <= 0.0)
